+++ zipkin4net-aspnetcore/Criteo.Profiling.Tracing.Middleware/project.json
@@ -31,7 +31,8 @@
   "dependencies": {
     "Microsoft.AspNetCore.Hosting": "1.1.1",
     "Microsoft.Extensions.Logging": "1.1.1",
-    "Criteo.Profiling.Tracing": "0.2"
+    "System.Net.Http": "4.3.0",
+    "Criteo.Profiling.Tracing": "0.3"
   },
   "frameworks": {
     "net451": {},
@@ -42,7 +43,7 @@
         ]
       },
       "dependencies": {
-        "NETStandard.Library": "1.6.1",
+        "NETStandard.Library": "1.6.1"
       }
     }
   }
