+++ zipkin4net/Criteo.Profiling.Tracing/Transport/ZipkinHttpTraceExtractor.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.Specialized;
-using System.Net.Http.Headers;
 
 namespace Criteo.Profiling.Tracing.Transport
 {
@@ -36,41 +35,7 @@ namespace Criteo.Profiling.Tracing.Transport
             return false;
         }
 
-        public bool TryExtract(HttpHeaders carrier, out Trace trace)
-        {
-            string encodedTraceId, encodedSpanId;
-
-
-            if (TryGetSingleValue(carrier, ZipkinHttpHeaders.TraceId, out encodedTraceId)
-                && TryGetSingleValue(carrier, ZipkinHttpHeaders.SpanId, out encodedSpanId))
-            {
-                string flagsStr, sampledStr, encodedParentSpanId;
-                TryGetSingleValue(carrier, ZipkinHttpHeaders.Flags, out flagsStr);
-                TryGetSingleValue(carrier, ZipkinHttpHeaders.Sampled, out sampledStr);
-                TryGetSingleValue(carrier, ZipkinHttpHeaders.ParentSpanId, out encodedParentSpanId);
-
-                return TryParseTrace(encodedTraceId, encodedSpanId, encodedParentSpanId, sampledStr, flagsStr, out trace);
-            }
-
-            trace = default(Trace);
-            return false;
-        }
-
-        private static bool TryGetSingleValue(HttpHeaders headers, string header, out string value)
-        {
-            IEnumerable<string> values;
-            bool success = headers.TryGetValues(header, out values);
-            var enumerator = values.GetEnumerator();
-            if (!enumerator.MoveNext())
-            {
-                value = default(string);
-                return false;
-            }
-            value = enumerator.Current;
-            return success;
-        }
-
-        internal static bool TryParseTrace(string encodedTraceId, string encodedSpanId, string encodedParentSpanId, string sampledStr, string flagsStr, out Trace trace)
+        public static bool TryParseTrace(string encodedTraceId, string encodedSpanId, string encodedParentSpanId, string sampledStr, string flagsStr, out Trace trace)
         {
             if (string.IsNullOrWhiteSpace(encodedTraceId)
                 || string.IsNullOrWhiteSpace(encodedSpanId))
